Externalize and test the data manipulation in ExploreRegionDropdown component
